requirements.txt
aiogram==3.4.1
gspread==6.0.0
google-auth==2.34.0
apscheduler==3.10.4
